+++ .eslintrc.cjs
@@ -2,5 +2,6 @@
 
 module.exports = {
   // FIXME: ./node_modules/ prefix is required, why?
-  extends: ['./node_modules/@julian_cataldo/astro-configs/.eslintrc.cjs'],
+  root: true,
+  extends: ['./node_modules/@julian_cataldo/astro-configs/eslint-all.cjs'],
 };