+++ .stylelintrc.cjs
@@ -1,5 +1,5 @@
 /** @type {import("@types/stylelint").Options} */
 
 module.exports = {
-  extends: ['@julian_cataldo/astro-configs/.stylelintrc.cjs'],
+  extends: ['@julian_cataldo/astro-configs/stylelint-all.cjs'],
 };