+++ app/Base/index.js
@@ -1,3 +1,9 @@
-import Base from './Base.astro';
+/* ——————————————————————————————————————————————————————————————————————————— *
+ *              © Julian Cataldo — https://www.juliancataldo.com.              *
+ *                      See LICENSE in the project root.                       *
+/* —————————————————————————————————————————————————————————————————————————— */
 
-export default Base;
+import Body from './Body.astro';
+import Document from './Document.astro';
+
+export { Body, Document };