+++ app/Breakpoints/index.js
@@ -1,0 +1,8 @@
+/* ——————————————————————————————————————————————————————————————————————————— *
+ *              © Julian Cataldo — https://www.juliancataldo.com.              *
+ *                      See LICENSE in the project root.                       *
+/* —————————————————————————————————————————————————————————————————————————— */
+
+import Base from './Breakpoints.astro';
+
+export default Base;