+++ app/ColorMode/index.js
@@ -1,3 +1,8 @@
+/* ——————————————————————————————————————————————————————————————————————————— *
+ *              © Julian Cataldo — https://www.juliancataldo.com.              *
+ *                      See LICENSE in the project root.                       *
+/* —————————————————————————————————————————————————————————————————————————— */
+
 import ColorMode from './ColorMode.astro';
 
 export default ColorMode;