+++ app/Prefetch/index.js
@@ -1,3 +1,8 @@
+/* ——————————————————————————————————————————————————————————————————————————— *
+ *              © Julian Cataldo — https://www.juliancataldo.com.              *
+ *                      See LICENSE in the project root.                       *
+/* —————————————————————————————————————————————————————————————————————————— */
+
 import Prefetch from './Prefetch.astro';
 
 export default Prefetch;