+++ app/ScrollObserver/index.js
@@ -1,3 +1,8 @@
+/* ——————————————————————————————————————————————————————————————————————————— *
+ *              © Julian Cataldo — https://www.juliancataldo.com.              *
+ *                      See LICENSE in the project root.                       *
+/* —————————————————————————————————————————————————————————————————————————— */
+
 import ScrollObserver from './ScrollObserver.astro';
 
 export default ScrollObserver;