+++ app/Transition/index.js
@@ -1,3 +1,8 @@
-import TransitionOverlay from './Overlay/TransitionOverlay.astro';
+/* ——————————————————————————————————————————————————————————————————————————— *
+ *              © Julian Cataldo — https://www.juliancataldo.com.              *
+ *                      See LICENSE in the project root.                       *
+/* —————————————————————————————————————————————————————————————————————————— */
+
+import TransitionOverlay from './Transition.astro';
 
 export default TransitionOverlay;