+++ configs/index.mjs
@@ -1,0 +1,3 @@
+// Stub
+
+export default function index() {}