+++ configs/prettier-base.cjs
@@ -1,0 +1,23 @@
+/** @type {import("@types/prettier").Options} */
+
+module.exports = {
+  printWidth: 80,
+  semi: true,
+  singleQuote: true,
+  tabWidth: 2,
+  trailingComma: 'all',
+  useTabs: false,
+};
+
+// {
+//   files: ['.*', '*.json', '*.md', '*.toml', '*.yml'],
+//   options: {
+//     useTabs: false,
+//   },
+// },
+// {
+//   files: ['**/*.css', '**/*.scss'],
+//   options: {
+//     singleQuote: false,
+//   },
+// },