+++ app/Prefetch/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.3](https://github.com/JulianCataldo/web-garden/compare/@julian_cataldo/astro-prefetch@0.12.2...@julian_cataldo/astro-prefetch@0.12.3) (2022-08-20)
+
+**Note:** Version bump only for package @julian_cataldo/astro-prefetch
+
+
+
+
+
 ## [0.12.2](https://github.com/JulianCataldo/web-garden/compare/@julian_cataldo/astro-prefetch@0.12.1...@julian_cataldo/astro-prefetch@0.12.2) (2022-07-31)
 
 **Note:** Version bump only for package @julian_cataldo/astro-prefetch