+++ components/LicensesReport/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.1](https://github.com/JulianCataldo/web-garden/compare/@julian_cataldo/astro-licenses-report@0.14.0...@julian_cataldo/astro-licenses-report@0.14.1) (2022-08-20)
+
+**Note:** Version bump only for package @julian_cataldo/astro-licenses-report
+
+
+
+
+
 ## [0.14.0](https://github.com/JulianCataldo/web-garden/compare/@julian_cataldo/astro-licenses-report@0.13.3...@julian_cataldo/astro-licenses-report@0.14.0) (2022-08-06)