+++ components/MediaMap/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.1](https://github.com/JulianCataldo/web-garden/compare/@julian_cataldo/astro-media-map@0.12.0...@julian_cataldo/astro-media-map@0.12.1) (2022-08-22)
+
+**Note:** Version bump only for package @julian_cataldo/astro-media-map
+
+
+
+
+
+## [0.12.0](https://github.com/JulianCataldo/web-garden/compare/@julian_cataldo/astro-media-map@0.11.5...@julian_cataldo/astro-media-map@0.12.0) (2022-08-21)
+
+
+### Features
+
+* support multiple leaflet map embed per page ([295ea80](https://github.com/JulianCataldo/web-garden/commit/295ea80812f39c801580d209957b37be84b568bd))
+
+
+### Bug Fixes
+
+* remove unused id attribute ([9a528cc](https://github.com/JulianCataldo/web-garden/commit/9a528cc38e856efe72a21fac0608e4430e6efcc0))
+
+
+
+## [0.11.5](https://github.com/JulianCataldo/web-garden/compare/@julian_cataldo/astro-media-map@0.11.4...@julian_cataldo/astro-media-map@0.11.5) (2022-08-20)
+
+**Note:** Version bump only for package @julian_cataldo/astro-media-map
+
+
+
+
+
 ## [0.11.4](https://github.com/JulianCataldo/web-garden/compare/@julian_cataldo/astro-media-map@0.11.3...@julian_cataldo/astro-media-map@0.11.4) (2022-07-31)
 
 **Note:** Version bump only for package @julian_cataldo/astro-media-map